refactor(data): expose operation lookups from moyo core
Move operations_from_number, operations_from_layer_number, and
magnetic_operations_from_uni_number from moyopy into moyo::data so
moyo-wasm and moyoc can reuse them without duplicating the logic.

The moyopy #[pyfunction] wrappers shrink to thin shims that adapt
PySetting / PyLayerSetting and wrap the result in PyOperations /
PyMagneticOperations. Python API is unchanged.
